Make the build reproducible
Whilst working on the Reproducible Builds effort [0], I noticed that
guidata could not be built reproducibly.

This is due to the testsuite not completely cleaning up after itself, so these
files get installed into the guidata Debian package. However, as they create
nondetermistic contents this makes the package unreproducible.

I originally filed this in Debian as bug #1041842 [1]. The patch is slightly
different there as I didn't notice the conditional block until now; this patch
should be prefered instead.

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/1041842